api/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
jinja2==3.1.4
python-multipart==0.0.9
SQLAlchemy==2.0.34
passlib[bcrypt]==1.7.4
python-jose==3.3.0
pydantic==2.10.3
pydantic-settings==2.6.1
requests==2.32.3
psycopg2-binary==2.9.9
